+++ src/lib/color-models/rgba.ts
@@ -1,10 +1,27 @@
+import { RGBA_REGEX } from '../../constants/regex'
 import { isRgba } from '../../types/isType'
-import { RGB, RGBA } from '../../types/types'
+import { HEX, RGB, RGBA } from '../../types/types'
+import { alphaToHex } from '../../utils/hex-utils'
 import { fromLongToShortRgbaFormat, shortRgbaFormatToRgbaObject } from '../../utils/rgba-utils'
+import { rgb2hex } from './rgb'
 
 // TODO: rgba2hex
 // TODO: rgba2cmyk
 // TODO: rgba2hsl
+
+/**
+ * Convert a rgba color object to a hex color.
+ * @param rgba color to convert to rgba
+ * @returns hex color
+ */
+export function rgbaToHex(rgba: RGBA): HEX {
+  if (!isRgba(rgba)) throw new Error(`${rgba} is not a rgba color.`)
+
+  const { r, g, b, a } = rgba
+  const rgbHex = rgb2hex({ r, g, b })
+  const alphaHex = alphaToHex(a)
+  return `${rgbHex}${alphaHex}`
+}
 
 /**
  * Convert a rgba color object to a rgb color removing the alpha value.
@@ -24,21 +41,17 @@ export function rgba2rgb(rgba: RGBA): RGB {
  * @param rgbaString rgba string
  * @returns rgba object
  */
-// TODO: check
-// TODO: do also for other color models
-// TODO: change regex to accept also a = .4
 export function rgbaString2Object(rgbaString: string): RGBA {
   // check short and long formats
-  const regexShortFormat = /^(([0-9]+)(\s)*,(\s)*([0-9]+)(\s)*,(\s)*([0-9]+)(\s)*,(\s)*((0(\.\d+)?|1(\.0+)?)))/gi
-  const regexLongFormat = /^((rgba(\s)*\()(\s)*([0-9]+)(\s)*,(\s)*([0-9]+)(\s)*,(\s)*([0-9]+)(\s)*, (\s)*((0(\.\d+)?|1(\.0+)?))(\s)*(\)))/gi
-  const isShortFormat = regexShortFormat.test(rgbaString)
-  const isLongFormat = regexLongFormat.test(rgbaString)
+  const isShortFormat = RGBA_REGEX.short.test(rgbaString)
+  const isLongFormat = RGBA_REGEX.long.test(rgbaString)
 
   if (!isShortFormat && !isLongFormat)
     throw new Error(
       `${rgbaString} is not a valid format. The accepted formats are 'r, g, b, a' and 'rgba(r, g, b, a)' with r, g, b in [0, 255] and a in [0, 1].`
     )
 
+  // convert rgbString to short format: 'R, G, B'
   const rgbaStringCleanShortFormat = isShortFormat
     ? rgbaString
     : fromLongToShortRgbaFormat(rgbaString)